+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#763](https://github.com/alleslabs/celatone-frontend/pull/763) Proposal status summary
+- [#764](https://github.com/alleslabs/celatone-frontend/pull/764) api v1 - proposal params
+- [#759](https://github.com/alleslabs/celatone-frontend/pull/759) Add proposal description and metadata
 - [#762](https://github.com/alleslabs/celatone-frontend/pull/762) api v1 - proposal validator votes info
 - [#745](https://github.com/alleslabs/celatone-frontend/pull/745) Add proposal top
 - [#758](https://github.com/alleslabs/celatone-frontend/pull/758) api v1 - proposal validator votes
@@ -56,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#761](https://github.com/alleslabs/celatone-frontend/pull/761) Fix copy button tooltip alignment
 - [#754](https://github.com/alleslabs/celatone-frontend/pull/754) Fix mobile guard incorrect behavior
 - [#751](https://github.com/alleslabs/celatone-frontend/pull/751) Fix fail txs should have no logs and remove stone-12-1